+++ zebra-state/src/service/read.rs
@@ -43,8 +43,8 @@ pub use find::{
     non_finalized_state_contains_block_hash, tip, tip_height, tip_with_value_balance,
 };
 pub use tree::{
-    ironwood_subtrees, ironwood_tree, orchard_subtrees, orchard_tree, sapling_subtrees,
-    sapling_tree,
+    history_tree, ironwood_subtrees, ironwood_tree, orchard_subtrees, orchard_tree,
+    sapling_subtrees, sapling_tree,
 };
 
 #[cfg(any(test, feature = "proptest-impl"))]